Refactor tests.

+++ skfda/tests/test_interpolation.py
@@ -401,7 +401,7 @@ class TestEvaluationSplineArbitraryImage(unittest.TestCase):
 @np.vectorize
 def _coordinate_function(
     *args: float,
-) -> np.typing.NDArray[float]:
+) -> np.typing.NDArray[np.float_]:
     _, *domain_indexes, _ = args
     return np.sum(domain_indexes)
